requirements.txt
Pillow
tensorflow
streamlit
numpy